requirements.txt
flask
flask-cors
typer[all]
requests
beautifulsoup4
lxml
Pillow
colorthief
Jinja2
weasyprint
markdown-it-py
pydantic
python-dotenv
openai
google-generativeai
pdfminer.six
python-docx
playwright>=1.46.0,<2
gunicorn 